refactor(llm-agent): move ILogger interface from llm-agent-server
